pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SANTIQ"
version = "0.1.0"
description = "A lightweight, modular, plugin-first ETL platform"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Dhritikrishna Tripathi", email = "dhritikrishnat@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "packaging>=22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "psutil>=5.9.0",
    "safety>=2.3.0",
    "pip-check-updates>=0.5.0",
]

[project.scripts]
santiq = "santiq.cli.main:app"

[project.entry-points."santiq.extractors"]
csv_extractor = "santiq.plugins.extractors.csv_extractor:CSVExtractor"

[project.entry-points."santiq.profilers"]
basic_profiler = "santiq.plugins.profilers.basic_profiler:BasicProfiler"

[project.entry-points."santiq.transformers"]
basic_cleaner = "santiq.plugins.transformers.basic_cleaner:BasicCleaner"

[project.entry-points."santiq.loaders"]
csv_loader = "santiq.plugins.loaders.csv_loader:CSVLoader"

[tool.setuptools.packages.find]
include = ["santiq*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["santiq"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = true
fail_under = 40

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers", 
    "--strict-config",
    "--cov=santiq",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests", 
    "compatibility: marks tests as plugin compatibility tests",
    "cli: marks tests as CLI-related tests",
    "external_plugin: marks tests for external plugin development",
    "github_workflow: marks tests specifically for GitHub Actions",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning", 
    "ignore::PendingDeprecationWarning",
]